refactor: split the application creation from sever creation
